Add remove deps cycle
